[refactor] product domain 폴더 계층 정리
