+++ src/main/java/com/nuclear/boomm/product/domain/Compliance.java
@@ -1,4 +1,4 @@
-package com.nuclear.boomm.product.domain.compliance;
+package com.nuclear.boomm.product.domain;
 
 import com.nuclear.boomm.common.BaseEntity;
 import jakarta.persistence.Column;
