+++ src/main/java/com/nuclear/boomm/product/domain/Product.java
@@ -1,4 +1,4 @@
-package com.nuclear.boomm.product.domain.product;
+package com.nuclear.boomm.product.domain;
 
 import com.nuclear.boomm.common.BaseEntity;
 import jakarta.persistence.Column;
@@ -11,8 +11,6 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Getter
