+++ src/main/java/com/nuclear/boomm/product/domain/RateBaseEntity.java
@@ -1,4 +1,4 @@
-package com.nuclear.boomm.product.domain.rate;
+package com.nuclear.boomm.product.domain;
 
 import com.nuclear.boomm.common.BaseEntity;
 import jakarta.persistence.MappedSuperclass;
