+++ src/main/java/com/nuclear/boomm/product/domain/SystemAndRegulationPrep.java
@@ -1,4 +1,4 @@
-package com.nuclear.boomm.product.domain.systemandregulationprep;
+package com.nuclear.boomm.product.domain;
 
 import com.nuclear.boomm.common.BaseEntity;
 import jakarta.persistence.Column;
